refactor(feature/routes): adjust components name and imports
